+++ src/es_utils/functions.rs
@@ -139,7 +139,7 @@ pub fn value_is_function(context: *mut JSContext, val: JSVal) -> bool {
 }
 
 /// check whether an object is a function
-pub fn object_is_function(cx: *mut JSContext, obj: *mut JSObject) -> bool {
+pub fn object_is_function(obj: *mut JSObject) -> bool {
     unsafe { JS_ObjectIsFunction(obj) }
 }